+++ src/Pwa/PageResult/Product/ProductPageResultHydrator.php
@@ -5,6 +5,7 @@ namespace SwagShopwarePwa\Pwa\PageResult\Product;
 use Shopware\Core\Content\Product\SalesChannel\SalesChannelProductEntity;
 use Shopware\Core\Content\Property\PropertyGroupCollection;
 use SwagShopwarePwa\Pwa\PageLoader\Context\PageLoaderContext;
+use SwagShopwarePwa\Pwa\PageResult\AbstractPageResultHydrator;
 
 /**
  * This is a helper class which strips down fields in the response and assembles the product page result.
@@ -15,7 +16,7 @@ use SwagShopwarePwa\Pwa\PageLoader\Context\PageLoaderContext;
  *
  * @package SwagShopwarePwa\Pwa\PageResult\Product
  */
-class ProductPageResultHydrator
+class ProductPageResultHydrator extends AbstractPageResultHydrator
 {
     public function hydrate(PageLoaderContext $pageLoaderContext, SalesChannelProductEntity $product, ?PropertyGroupCollection $configurator): ProductPageResult
     {
@@ -24,6 +25,10 @@ class ProductPageResultHydrator
         $pageResult->setProduct($product);
 
         $pageResult->setConfigurator($configurator);
+
+        if ($seoCategory = $product->getSeoCategory()) {
+            $pageResult->setBreadcrumb($this->getBreadcrumbs($seoCategory, $pageLoaderContext->getContext()));
+        }
 
         $pageResult->setResourceType($pageLoaderContext->getResourceType());
         $pageResult->setResourceIdentifier($pageLoaderContext->getResourceIdentifier());